+++ docs/release-notes.md
@@ -1,5 +1,22 @@
 # Release Notes
 
+## Unreleased
+
+### Added
+
+- Add lazy_setup parameter for testclient.
+- `disconnect` and `connect` return if they performed the setup/cleanup.
+
+### Fixed
+
+- Restore eager setup behavior in testclient.
+- Fix drop_dabase with postgresql.
+- Fix setup/disconnect hangups.
+
+### Removed
+
+- Dependency on nest_asyncio. It is an archived repo.
+
 ## 0.9.2
 
 ### Added